Fix backend of export messages
There are a few fixes:
(1) Convert the function to a class to follow the django-rest-framework permissions process. Using `login_requried` will try to redirect to a non-existent route
(2) Convert the streaming response to a normal `HttpResponse`. Happy to go back to `StreamingHttpResponse` if you can get it working. SHR has the advantage of working with a very large dataset. As the code was written it wouldn't work using the `with` syntax because the `StringIO` stream gets closed too early. Even once that was 'fixed' I was getting blank CSV files.

To test this code:
Comment out the `permission_classes` line and navigate to `http://127.0.0.1:8000/api/messages.csv`