FIX: new messages now have a username
Why does the backend send with the `sender_name` property and the frontend expect/need the `username` property? Where else does this conversion happen? (since it worked previously when you clicked away and clicked back).

My guess is this is necessary to translate our data model to what the vue-advanced-chat expects.