FIX: Get the new messages divider working
This is quite a hack, but I tried to document the hack clearly in case someone wants to find a better way to do it. One option would be to just fork the chat component and add our own fixes, but maybe there will be better docs in the future and it'll be clear how to do things the 'right' way.